File is added for 1st phase of 'News Classification and Analysis Project'.
This is to bring to your notice, that we are starting with 1st phase of News classification and analysis project. First phase will consist of data collection tasks. Problem statement and sample data has been shared with you in the attachment. Kindly download it and understand the problem statement. In this phase we have to scrape news data from various news archives. You will be allocated task to scrape archives from different time frames and news sites. The tasks will be allocated to you individually. Check your tasks in PMT